+++ Html/socketSample.html
@@ -1,1 +1,8 @@
-<script src="Js/socketSample.js"></script>
+<head>
+    <meta charset="utf-8">
+</head>
+<body>
+    <div id="main"></div>
+    <input type="button" value="test" id="testBtn">
+    <script src="Js/socketSample.js"></script>
+</body>
